refactor(esp32): change UartLink packet count from uint8_t to uint16_t
- Update packetCount parameter type in updateDerivedVerticalSpeeds() from uint8_t to uint16_t
- Change lastPacketCount_ member variable type from uint8_t to uint16_t
- Aligns with protocol's actual packet count range and prevents overflow issues

+++ receiver/esp32-companion/src/serial/uart_link.cpp
@@ -11,7 +11,7 @@ UartLink::UartLink(HardwareSerial& serial, uint32_t baud, int rxPin, int txPin)
     : serial_(serial), baud_(baud), rxPin_(rxPin), txPin_(txPin) {}
 
 void UartLink::updateDerivedVerticalSpeeds(uint32_t sampleTms,
-                                           uint8_t packetCount,
+                                           uint16_t packetCount,
                                            CompanionState& ioState) {
   const bool hasNewPacket = !havePacketCount_ || packetCount != lastPacketCount_;
   if (!hasNewPacket) {
